medicore-main/client/requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
requests